# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Movement",
        "version": "0.1"
    },
    "paths": {
        "/api/movements/run": {
            "get": {
                "tags": [
                    "Movement"
                ],
                "summary": "Run the car",
                "description": "Run the car forward",
                "operationId": "run",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/movements/stop": {
            "get": {
                "tags": [
                    "Movement"
                ],
                "summary": "Stop the car",
                "description": "Stop the car",
                "operationId": "stop",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/movements/backward": {
            "get": {
                "tags": [
                    "Movement"
                ],
                "summary": "backward the car",
                "description": "backward the car",
                "operationId": "backward",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/movements/forward": {
            "get": {
                "tags": [
                    "Movement"
                ],
                "summary": "forward the car",
                "description": "forward the car",
                "operationId": "forward",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/movements/exit": {
            "get": {
                "tags": [
                    "Movement"
                ],
                "summary": "exit the car",
                "description": "exit the car",
                "operationId": "exit",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}